[PHP-Symfony] Fix handling of null date-time parameter
The deserializeString method of the generated class JsmSerializer returns null
for other types of string, but not for date-time, instead it returns a DateTime
which represents "now" (because that what `new DateTime(null)` does).

Consequently when an API declares a date-time parameter as non-required and
when that endpoint is called without that parameters, then the user code
would end up having "now" instead of "null" for this parameter.

This commit fixes it