first iteration to try to fix #16123
This commit fixes the code generated with this OAS:

    openapi: 3.0.3
    info:
     title: test
     version: v0.1
     description: test
    paths:
     /api/path:
      get:
       responses:
        '200':
          description: ok
          content:
           application/json:
            schema:
             $ref: '#/components/schemas/TimeZone'
    components:
      schemas:
       TimeZone:
        type: object
        properties:
          someField:
           type: array
           items:
             type: string

However it breaks the code generated with this one (which is handled
fine before this commit):

    openapi: 3.0.3
    info:
     title: test
     version: v0.1
     description: test
    paths:
     /api/path:
      get:
       responses:
        '200':
          description: ok
          content:
           application/json:
            schema:
             $ref: '#/components/schemas/TimeZone'
    components:
      schemas:
       TimeZone:
        type: object

because in that case {{returnType}} is just "object", and that
Org.OpenApiTools.Model.Object does not exist.

A way to go farther could be for instance to always generate entities.
An alternative could be to try to find a way in api.mustache to find out
if the namespace to apply should be "system" (nb: that would work for
"type: object" type as above, but also for "type: string" ones)